daysAgo and ZiptoCity moved to utils folder
